DNA-623 JDBC Metadata TCK Test Failing in Nightly Integration Build
Now that the query tests are enabled, the TCK test is running XPathPosIndexTest against all of the connectors.  This test has a written prerequisite (in Javadoc) that the root node have at least 3 child nodes named the same as the value of the "nodeName1" property.  However, the test checks for a different precondition (more than three child nodes in total).

Since the JDBC Metadata connector doesn't support same-name siblings, it cannot meet the written prerequisite, but the configured root node did have more than 3 child nodes.  So the test tried to execute, but the query (which ended in [2]) couldn't possibly return values from this connector.

The applied patch moves the root for the test cases to a node that only has two children, effectively bypassing this test[1].  This is actually better for most other tests since it moves the test root closer to the actual root of the repository and exposes more nodes to TCK testing.  I tried to do this originally when I added the connector, but the performance was too slow.  

[1] - This is why the FileSystemRepositoryTCKTest didn't also break.  Although it doesn't support SNS either, the root node happened to only have two children so this test was effectively bypassed for that connector as well.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1532 76366958-4244-0410-ad5e-bbfabb93f86b