DNA-618 SVN connector integration tests are failing with '404 Not Found'
Applied patch that shrinks the size of the repo, making it run dramatically more quickly from sites outside of the jboss.org network. Prior to the search code, there was no particular penalty for having a huge repo as long as you didn't try to read every node. But the search code has to read the whole repo to build the search index. In this case, that required reading several thousand files, each requiring two network calls (one for the file and one for its contents). 

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1541 76366958-4244-0410-ad5e-bbfabb93f86b