DNA-616 The problem has to do with the fact that this example uses a FederatedRepositorySource as a source for a JpaRepository (which creates a FederatedRepositorySource on top of the supplied source).  Thus, there is a FederatedRepositorySource that uses another FederatedRepositorySource, and this is the condition that causes the deadlock.
When the JcrRepository is created, it uses a SearchIndexer to crawl the content and create the requisite indexes for query/search.  The SearchIndexer is create a CompositeRequestChannel (which it uses to submit to the designated repository source a CompositeRequest that remains open, allowing the SearchIndexer to supply requests and then use those requests to build other requests that are submitted within the same execution/transaction), and is submitting a ReadBranchRequest to the source that happens to be the lower FederatedRepositorySource. The FederatedRepositorySource returns a connection that has an execute(...) method that implements the work by using the fork-join algorithm to process the request, which in this case is the  CompositeRequest (owned by the CompositeRequestChannel created by the SearchIndexer).

Now, the ForkRequestProcessor works on all the incoming request, which again is the CompositeRequest.  The FRP's process(CompositeRequest) is actually inherited from RequestProcessor, and simply uses an iterator to get all of the requests within the CompositeRequest.  The iterator promptly returns the first ReadBranchRequest, so the ForkRequestProcessor handles this via it's process(ReadBranchRequest), and like all of the ForkRequestProcessor.process(...) methods, forks the request into appropriate requests to submit to the underlying (federated) sources, and then it enqueues the forked requests into a queue.

Note that the ForkRequestProcessor does NOT mark the requests it processes as being completed.  After all, this is the job of the JoinRequestProcessor.

Meanwhile, the ForkRequestProcessor.process(ReadBranchRequest) completes and returns, and the iterator in the ForkRequestProcessor.process(CompositeRequest) then attempts to get the next request.

But the SearchIndexer has only added one ReadBranchRequest and is waiting/blocking until that request is completed before it continues.  But the FederatedRepositoryConnection is waiting/blocking until the next request is added to the CompositeRequest.  Thus the deadlock.

The only way to break this deadlock is for the ReadRequestBranch to be completed.  But, because of the way that FederatedRepositoryConnection.execute(...) is implemented, the ForkRequestProcessor.process(request) must complete before the JoinRequestProcessor is started.  So we need to create and start the JoinRequestProcessor without waiting for the ForkRequestProcessor.process(request) to complete.

The solution is actually coded in the FederatedRepositoryConnection.execute(...) - there is a 'shouldProcessSynchronously(Request)' method that returns whether the ForkRequestProcessor should be processed synchronously or asynchronously, and is implemented to always return 'true' (synchronous).  The solution to this problem requires that the method return 'false' for CompositeRequests that contain more than one (or an unknown number) of requests.  (In other cases, there is just one request, so the forking can be done synchronously, saving the work of pushing the fork processing into the executor service.)

Thus, the 'shouldProcessSynchronously(Request)' was implemented to do this, and the test completed ... with an error.  This error is caused by a bug in the JoinRequestProcessor.process(ReadBranchRequest) that failed to properly calculate the projected actual Locations for the ProxyNodes from the source requests.  (Apparently the SearchIndexer reads a branch that is deeper than was previously tested.)

Fixing the JoinRequestProcessor.process(ReadBranchRequest) allows the repository example test cases to complete successfully.

See the attached patch, which was committed and verified with a full integration build (all unit and integration tests pass).  The patch also changes the log level of the repository example back to 'ERROR' (what it was previously), and also re-enables the 'docs/examples/gettingstarted/repository' project in the 'docs/examples/gettingstarted/pom.xml'.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1547 76366958-4244-0410-ad5e-bbfabb93f86b