DNA-617 NPE while updating indexes (during integration tests)
Attach patch that corrects the behavior of the BasicJpaConnector with regard to EntityManager lifetime.

Prior to the patch, the BJC obtained an EntityManager in its constructor and released the EntityManager in its close method.  The typical lifecycle for a connection is a call to the constructor followed by a single call to execute(...) followed by a call to close().  With this lifecycle, the EntityManager is only used for one call to execute(...) and the transaction is committed or rolled back after the call as appropriate.

Now that the RepositoryConnectionPool is being used (and was fixed to reuse connections), the typical lifecycle is one call to the constructor followed by multiple calls to execute(...) within different transactions followed by (eventually) a call to close().  Since the same EntityManager was being held for the entire life of the BJC object, the EntityManager would be used for multiple transactions without any attempt being made to clear out its cache.  We're using Hibernate as our JPA implementation, so the Hibernate session cache was gradually accumulating copies of the data in the database _that were gradually falling out of sync with the database_.

In other words, BJC1 would be used to create a node and set a property leaving the ChildEntity and the PropertiesEntity for the node in the Hibernate session cache for BJC1's EntityManager.  BJC2 would then be used to read the same node and update its property.  Since the entities for the node weren't yet in in the session cache for BJC2's EntityManager, BJC2 would execute select statements to load the necessary data and would perform the updates.  THEN in a subsequent call, BJC1 would be reused to read from the same node.  Since BJC already had the ChildEntity and PropertiesEntity for the node in its session cache, it would reuse the cached (and now stale) versions instead of trying to hit the database.  Hilarity ensued.

The patch makes BJC obtain an EntityManager at the beginning of the execute(...) call and release the EntityManager at the end of the execute(...) call.  Since "released" EntityManager objects are really closed by the EntityManagers factory object, this will cure the session cache issue.

This fixes all but 5 of the TCK failures with the JPA connector using the basic model.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1578 76366958-4244-0410-ad5e-bbfabb93f86b