DNA-644 BasicJpaRepositoryTckTest has failures related to XPath queries
Applied patch that makes BasicRequestProcessor.process(CreateNodeRequest) call recordChange(request) after successful node additions. If recordChange(...) isn't called, the SearchEngineIndexer is never made aware of the change and the node will not be added to the indexes or returned as a search result. Adding this line makes all 5 failing tests pass. I reviewed the rest of the class, but could not find any other places where this call was missing.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1590 76366958-4244-0410-ad5e-bbfabb93f86b