MODE-631 Provide a way to set the JcrRepository descriptors from the JcrConfiguration and JcrEngine
Applied patch that looks for a mode:descriptors tag in each mode:repository. Children of this tag that have a name and value attribute (regardless of the namespace) are treated as a descriptor entry and passed to the constructor for the corresponding JcrRepository. This is analogous to how options are handled. The patch also expands one of the test cases to verify that this is working.


git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1617 76366958-4244-0410-ad5e-bbfabb93f86b