MODE-642 Added a 'setGrammars(String[])' method that can be used to specify which grammars should be used during sequencing. The values can be the name of the grammar (e.g., the value of the DdlParser.getId() method) for any of the built-in parsers, or the full-qualified name of the DdlParser implementation class. If not set, or if set with a null or empty array, the built-in grammars will be used.
Also added a 'setClasspath(String[])' method that can be used to specify the names of the classloaders that should be used to load the DdlParser implementation class. If not set, or if set with a null or empty array, the default classloader will be used.

Added some integration tests that verify this behavior, and also fixed a few minor issues in the SequencerConfig and StreamSequencerAdapter classes around calling these JavaBean-style setters when the parameter type is an array of String objects (rather than an Object[] with String values).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1640 76366958-4244-0410-ad5e-bbfabb93f86b