MODE-636 JcrConfiguration files should be able to specify node types by referencing one or more CND files
Applied patch that allows the mode:resources property of the jcr:nodeTypes node in the configuration repository to provide a comma-delimited list of paths to CND files in the classpath. Adding this capability allows for some minor simplification of the ModeShapeRepositoryStub, which, in turn allows the various TCK tests to be used as test cases for this new functionality.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1655 76366958-4244-0410-ad5e-bbfabb93f86b