MODE-670 The modeshape-jcr-tck test case classes were being run in an arbitrary order (however JUnit/Surefire determines them), and the Level 1 TCK unit tests were being run after some of the other TCK unit tests that modified the repository.  When some of those tests fail, apparently the repository is not being returned to its original state, and thus there is missing content required by the Level 1 TCK tests (which are all read-only).  The fix was to add a static method to ModeShapeRepositoryStub that will force the JcrEngine to be rebuilt, and then to add a test case to each of the test suites that calls this method.
Normally, the org.apache.jackrabbit.test.JCRTestSuite runs the Level 1 tests first, followed by the Level 2 tests, the query tests, the node type tests, the lock tests, the version tests, and then the observation tests.  So the order that the modeshape-jcr-tck tests are run does not reflect the actual TCK test order, so forcing the recreation of the engine IMO works fine.

Really, the 'correct' way to fix this is to create a single TestSuite that adds the appropriate tests in the appropriate order (e.g., org.apache.jackrabbit.test.JCRTestSuite class). However, we want to keep them separated and grouped by feature so that it is more obvious which test failures are for which (optional) features.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1668 76366958-4244-0410-ad5e-bbfabb93f86b