MODE-679 Attempting to read the MANIFEST.MF doesn't work well, because it grabs the first 'META-INF/MANIFEST.MF' file on the classpath (whatever that is). Therefore, I decided on the approach of using a specific file ('modeshape-jcr/src/main/resources/org/modeshape/jcr/repository.properties') and automatically updating that file with Maven.
This works better, but may require running a 'mvn compile' from the command line before the tests in JcrRepositoryTest can be run from within Eclipse (because Maven inserts the POM properties into this file during the compilation phase, whereas compiling within Eclipse will not do this).  Hopefully this will not cause a problem.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1694 76366958-4244-0410-ad5e-bbfabb93f86b