MODE-686 Added to JcrQueryManagerTest several test cases that verified this behavior.  These queries were converted to JCR-SQL2 queries (actually, internal abstract query model queries that can be expressed exactly in JCR-SQL2) with a path constraint such as the following:
   PATH(nodeSet) LIKE '%/NodeA'

However, this LIKE expression with a leading '%' was getting converted incorrectly to a Lucene query (essentially, '/%/NodeA', which of course never matched nodes that appeared directly under the root).

Once this was fixed, the new test cases all passed.

Also discovered a similar issue when the path literal contained same-name-siblings. In this case, the conversion to a Lucene RegexQuery neglected to properly escape the '[' character (which is a meta-character in regular expressions). Additional test cases were added to cover this particular condition.

A full integration build passed with these fixes.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1696 76366958-4244-0410-ad5e-bbfabb93f86b