MODE-691 I first added an integration test that replicates this same condition, and verified it fails for MySQL.
The issue was indeed an HQL query that used a select (with a join) in the criteria of a delete statement. Since this is not supported by MySQL, I changed the way that this delete statement was being used. For dialects other than MySQL, the behavior is the same.

However, for the MySQL dialects (actually, any dialect in lowercase form that contains 'mysql'), the hashes for the unused large entity values are found by a query and then one or more delete statements are used where each delete statement uses an 'IN' criteria to specify the hash and where no more than 20 values for the 'IN' criteria are used. In other words, the unused large values are removed in sets of 20 (which is still much better than removing them individually).

The test cases work for MySQL dialect, and no other regressions for other DBMSes were found.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1706 76366958-4244-0410-ad5e-bbfabb93f86b