MODE-690 I started to fully implement this method, but it really does require a fair amount of changes (mostly additions) to SessionCache and GraphSession. Plus, a lot of the code involves validating the state of the as-of-yet-unsaved changes elsewhere in the session, such as ensuring that the changed property doesn't contain references to nodes that were created elsewhere in the session, that the primary type and/or mixin types have not changed (lest the property validation for the property being saved cannot be done according to the persisted state), and that the property was changed from a multi-valued to a single-valued property (or vice versa) resulting in the need to also change the internal property used to track multi-valued properties (which is even more complicated if other properties are also changed in a similar manner and thus the internal property must be saved in a partial form with just those changes required by the property being saved).
Taking all this into account, plus the fact that JCR 2.0 no longer defines the Item.save() functionality and that this method is deprecated in the JCR 2.0 Java API, led me to conclude that the best way at this time to implement this is to simply delegate to parent Item.save().  So this is what was done.

I would also be surprised if very many applications use Property.save(), as it is a rather odd notion to be partially updating a single property. Instead, I expect that Node.save() is used quite a bit more often. Also, I can imagine where this might be needed within a JCR implementation so that internal components can persist some of the changes to its own repository content (e.g., namespace manager, authorization manager, etc). However, ModeShape uses the lower-level Graph API to immediately manipulate and persist a subset of the content independently of the JCR Session.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1714 76366958-4244-0410-ad5e-bbfabb93f86b