MODE-281 Support accessing repository through WebDAV
Applied a patch that provides a WebDAV implementation for ModeShape (although there aren't any ModeShape dependencies in there yet - it could be used with any JCR implementation).  The patch splits out some of modeshape-web-jcr-rest into a separate modeshape-web-jcr project.

The patch uses the SF.net WebDAV library suggested by Randall (thanks!) as the backbone of the implementation, albeit modified with a couple of OS X workarounds that the Drools Guvnor team laid on top of the same library for their WebDAV implementation.  These are noted in the ModeShapeWebdavStore.java.

This approach maps every node that has a primary type of nt.:file (or one of its descendants) to a WebDAV resource and every other node to a WebDAV folder.  That allows for browsing through existing hierarchies, but does not fully exposed the data in non-file nodes.  Just as with the REST server, authentication information is proxied in from the Servlet engine, allowing for integrated security.

The default implementation requires a repository and workspace name to be provided in the web.xml, but there are a pair of interfaces that can be implemented (RequestResolver and UriResolver) to provide extremely flexible mappings if that is desired in the future.

The implementation also aims to be configurable wherever reasonably possible, including (but not limited to):
- the choice of RequestResolver/UriResolver implementation
- the node type to use for new folders, files, and content nodes
- the node types to treat as a folders, files, or content nodes for existing nodes

One hardwired parameter that may need to be configurable in the future is the use of the Aperture MIME type detector.  I really didn't want to tightly couple the WebDAV implementation to ModeShapeEngine (or JcrEngine) just to reuse the MIME type detectors there, so I just hard-coded that.  If anyone wants me to change, I'll happily do so as a Good Practice.

The committed patch differs slightly from the patch in JIRA, as it removes spurious .settings directories and cleans up the .classpath entries in Eclipse to conform to our standards.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1724 76366958-4244-0410-ad5e-bbfabb93f86b