MODE-713 Recently Registered Namespace Mappings Aren't Respected by the Query Engine
The JcrRepository was passing executionContext instead of this.executionContext as an argument to the query engines.  This caused the query managers and the repository to have different execution contexts with different namespace registries.  Namespace mappings registered through the workspace would not be reflected in the query manager's namespace registry.  This patch corrects that and adds a test case to verify the behavior.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1754 76366958-4244-0410-ad5e-bbfabb93f86b