MODE-720 Can't Set Reference Property to Unsaved Node if Constraint Exists on Property
Applied patch that first checks through the nodes already loaded by this session when attempting to load a node by UUID (as is done by the constraint checker for reference nodes, among other places).  If the node with the given UUID hasn't already been loaded into the session, THEN the SessionGraph attempts to read the node from the graph and load it into the session.  The patch also adds a test case to verify this behavior.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1760 76366958-4244-0410-ad5e-bbfabb93f86b