MODE-701 Autocreated Child Nodes and Properties Aren't Autocreated
Applied patch that refactors SessionCache.addAutoCreatedItems so that it is more difficult to give it
the wrong NodeEditor to check.  The patch also adds logic to JcrContentHandler to ensure that autocreated items are imported correctly and fixes a minor issue with AbstractJcrNode.getNode(Name).  Finally, the patch adds a test case to help verify its correctness.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1765 76366958-4244-0410-ad5e-bbfabb93f86b