MODE-727 Allow FileSystemSource to set the CustomPropertyFactory class in configuration
Applied patch that overloads setCustomPropertiesFactory to take the class name of the factory as well as an instance of the factory.  The patch also modifies the getReference/getObjectInstance to put the class name of the factory in the Reference.  This is consistent with requirement that the CustomPropertiesFactory be immutable.  All tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1766 76366958-4244-0410-ad5e-bbfabb93f86b