MODE-730 Make XMLSchema Namespace Prefix Conform to JackRabbit
Applied patch that changes the default prefix for the XMLSchema namespace from xsd to xs.  This should not cause any repositories that map the jcr:system projection to a persistent store.  However, it may require changes to any repositories that hardcode access to properties or nodes that utilize this namespace.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1771 76366958-4244-0410-ad5e-bbfabb93f86b