MODE-732 GraphSession.findNodeRelativeTo Throws Different Exception for Deep Missing Node
This patch catches the RequestException that gets thrown if more than one segment in the relative path doesn't exist and attempts to read the relative path one node at a time. The first non-existing node causes a PNFE to be thrown that accurately reflects both the intended target of the method and the lowest existing path.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1772 76366958-4244-0410-ad5e-bbfabb93f86b