MODE-736 JndiRepositoryFactory Has a configFile Property That Points to a Resource - Not a File
Applied patch that makes the factory first try to read the value of the configFile parameter as a resource on the classpath.  If this fails, it then tries to read the value as the name of or path to a file.  The patch also adds a test case to verify this behavior.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1786 76366958-4244-0410-ad5e-bbfabb93f86b