MODE-739 Add Support for Node.getProperties(String[]) and Node.getNodes(String[])
Applied patch that adds the new getNodes and getProperties variants while refactoring the old getNodes and getProperties methods to use the new methods.  The patch also adds some tests for the new and old versions of the methods.  This exposed a bug in the existing getProperties implementation, so that was fixed and made more consistent with the getNodes implementation.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1788 76366958-4244-0410-ad5e-bbfabb93f86b