MODE-714 Create a first-cut at the Map-based version of this. Rather than replace the existing connectors, this prototype was added in a new package, allowing it to be developed and tested side-by-side with the existing connectors. This also means that individual connectors can be migrated one-by-one, and eventually the old map connector (and path connector) can be deprecated and/or removed. Indeed, a new In Memory Repository implementation was added and currently passes most of the same unit tests as the older IMR, and this new IMR will eventually replace the older implementation.
Note that this new implementation is not currently as fast as the older Map connector, because any navigation or determination of paths must involve accessing the (parent or child) UUID from the MapNode followed by looking up the corresponding MapNode object first in the transaction's state or (if not found) in the workspace's map. The old connectors don't need to do this, since each MapNode has a direct reference to the parent and children (which is the cause of serialization difficulties). Additionally, the changes made within the transaction are pushed into the workspace's map only upon commit, so this is extra work the old connector is not performing.

The new design, however, should allow the new IMR, Infinispan, and JBoss Cache connectors to all be extremely simple, since the only difference really is how the Map<UUID,MapNode> is obtained/created. (There may be other differences as well, but only if the new connectors take advantage of additional Infinispan and JBC features currently not used by the old connectors.)

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1807 76366958-4244-0410-ad5e-bbfabb93f86b