MODE-714 Committed a new base connector for map-oriented repositories, and changed the InMemory and Infinispan connectors to make use of it. This new approach uses a transactional (i.e., 'all or nothing') technique to either commit all changes or to commit no changes if there are any errors in any of the requests. Plus, this approach uses a correct serialization technique, as none of the serialized nodes contain any information for a node other than its properties and list of UUIDs for its children.
New unit tests were added to use Infinispan with file-based persistence. These tests failed when running against the original Infinispan connector, but pass with the new version of the Infinispan connector that uses corrected serialization.

The JBoss Cache connector needs to be converted, as does the JPA Simple model. We also probably want to create a path-oriented base class and use it in the path-oriented connectors (e.g., FileSystem, SVN, etc.).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1810 76366958-4244-0410-ad5e-bbfabb93f86b