MODE-711 Allow InfinispanConnector(s) to create configurable Infinispan cachemanagers
Applied patch that will try to initialize a DefaultCacheManager from the given configuration name if the cacheConfigurationName property is present.  The InfinispanSource will first try to treat the cacheConfigurationName property as a pointer to a resource that can be loaded by the classloader.  If this fails, the cacheConfigurationName will be assumed to be a path to a file.  The patch also adds tests to verify this behavior.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1812 76366958-4244-0410-ad5e-bbfabb93f86b