MODE-744 Complete Support for JCR2 Node Type Features
Applied patch that implements almost all of the remainder of the JCR2 node type features.  Fixing MODE-752 enabled several tests that had previously been skipped that exposed more bugs.  At this point, the node type code is about as consistent as it can be with the JCR2 implementation.

The remaining piece to implement is validation of names in the template setters.  Unfortunately, to make this truly compatible would require the signature of the templates to be modified to throw ConstraintViolationExceptions, and I am trying not to modify methods or classes in ways that would break previously working code.  A separate defect for this will be opened.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1814 76366958-4244-0410-ad5e-bbfabb93f86b