MODE-733 Fixed a number of issues that were identified by the query-related JCR TCK unit tests, which are all passing at this point except for: - those tests that expect the '/jcr:system' node to be found by search/query (see MODE-760); - those tests using the 'text()' JCR-SQL function in criteria (not supported at this time; see MODE-759); and - those (few) tests that produce invalid join criteria in their test queries (e.g., 'SELECT * FROM nt:base, test:queryable WHERE test:queryable.jcr:path = mix:referenceable.jcr:path' in SQLJoinTest.testJoinNtBase)
The repository content used in the tests needed to be modified quite a bit, primarily because several of the test cases would fail when running the entire suite, but would succeed when running just the query tests. I concluded that these tests were expecting data that was being messed up in the tearDown methods of prior tests, so the easiest way to fix this was to just use separate branches in the content for these tests.

At this point, all of the included JCR TCK tests are running, and we only have 6 errors and 3 failures when running the complete TCK. Two of these failures are from outstanding observation issues, while the remainder are due to the three conditions outlined above.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1834 76366958-4244-0410-ad5e-bbfabb93f86b