MODE-764 Use Constant for Default Value of rootNodeUuid Property in InfinispanSource
Applied patch that changes the default value of the rootNodeUuid property to a value that parses as a UUID (but is obviously a default value).  The patch also creates test cases at the graph and JCR levels that test some very basic behavior in a cluster.  Finally, the patch adds a bit of documentation to the Infinispan connector in the Reference Guide to describe the requirements for using the InfinispanSource in a cluster.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1848 76366958-4244-0410-ad5e-bbfabb93f86b