MODE-762 Added checks to ensure the session is open when methods are called on our implementations of Session, Workspace, NodeTypeManager, NamespaceRegistry, Node, Property, VersionManager, QueryManager, and ObservationManager. While I'm not 100% sure that I've covered all of the public methods, I pretty sure I've covered all of the important public methods.
The checks basically delegate to a new JcrSession.checkLive() final method that simply throws an exception if the 'isLive()' method returns false. (The isLive() method is not final and can be stubbed in a mock object, and in fact a number of tests needed to be modified to stub the isLive() method.) As such, this check is pretty fast and efficient, although obviously calling it more times than necessary is not desirable.

To reduce the number of times the check is performed, I chose to place these checks in (most) public methods rather than in protected/private helper methods that may be used frequently within a single public method invocation. This most likely resulted in more explicit calls (and thus more code), but hopefully it means that fewer checks are performed in those public methods that makes multiple calls to internal helper methods.

Also, I tried to ensure that this check is not done in the few methods of Node and Property that might be called to help identify a node that belongs to a closed session. For example: getPath(), toString(), getName(), getUuid(), etc.

At this point, all unit and integration tests pass, and there doesn't appear to be a noticeable impact on performance (given the range of build times I have seen lately).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1849 76366958-4244-0410-ad5e-bbfabb93f86b