MODE-771 Remove Sequencer Dependency on Namespace Mappings
Applied a patch that deprecated SequencerOutput.setProperty(String, String, Object...) and SequencerOutput.setReferences(String, String, Object...) in accordance with the description of this issue.  The patch also converts all of the existing sequencers (and their tests) to the SequencerOutput.setProperty(Path, Name, Object...) method.  This method is the recommended replacement for the deprecated method.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1853 76366958-4244-0410-ad5e-bbfabb93f86b