MODE-765 Add Support for JCR2 Locking
Applied a patch that adds the extra locking functionality specified in JCR 2.  This largely entails minor changes to existing classes, although two new classes needed to be added to modeshape-jcr-api to adapt the JCR 1 locking classes to the JCR 2 API.  The only substantive new class is JcrLockManager, an implementation of the LockManager interface that provides methods for interfacing with locks.  As many of these duplicate existing methods in AbstractJcrNode and JcrSession, the existing implementations have been moved into the new class.

All tests still pass and the more extensive locking tests in JCR 2 exposed a bug in the AbstractJcrNode.holdsLock() method that has been fixed as well.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1855 76366958-4244-0410-ad5e-bbfabb93f86b