MODE-770 Add Support for JCR2 RepositoryFactory
Applied patch that provides a working implementation of the RepositoryFactory interface that looks for a JCR URL to determine which repository to return.  This version passes all JCR 1 tests, but has not yet been tested against the JCR 2 test suite.

The patch uses the URL format described in the most recent forum posts on the subject and adds shutdown methods in a ModeShape extension to the JCR 2 API.

This patch also updates the Getting Started Guide and the Reference Guide to document the new changes.  All tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1860 76366958-4244-0410-ad5e-bbfabb93f86b