MODE-770 Add Support for JCR2 RepositoryFactory
Applied patch that provides a working implementation of the RepositoryFactory interface that looks for a JCR URL to determine which repository to return.  This version passes all JCR 1 tests, but has not yet been tested against the JCR 2 test suite.

Additionally, the reference guide, getting started guide, and examples should be converted
to use this class as well.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1861 76366958-4244-0410-ad5e-bbfabb93f86b