MODE-776 Upgrade to JCR 2.0 API
Applied patch that changes the JCR API version from 1.0.1 to 2.0.  The patch stubs in many missing methods that have yet to be implemented and updates JcrTckTest to use the 2.0 tests.

The patch also removes the BasicJpaIntegrationTest and its related configuration as the BasicModel has been deprecated and is scheduled for removal prior to the 2.0 release.

All tests still pass (modulo the new JCR 2.0 tests that don't work yet and are commented out).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1862 76366958-4244-0410-ad5e-bbfabb93f86b