MODE-753 Add Immediate Validation to Node Type Management Classes
Applied patch that takes advantage of the new JCR2 support to add the missing exceptions to the node type template methods that allow immediate validation.  In the process of testing these defects, it also became apparent that a few other changes needed to be made to support the new property types.  This led to the addition of a factory for decimal properties in the default TypeSystem and minor cleanup in the setRequiredType method of JcrPropertyTemplate.

Additionally, the patch attempts to remove now-redundant reimplementations of JCR2 method signatures in the o.m.j.api and o.m.j.nodetype packages.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1867 76366958-4244-0410-ad5e-bbfabb93f86b